added confirm service to client

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core'
 import { Message } from '../models/message'
 import { Pagination } from '../models/pagination'
+import { ConfirmService } from '../_services/confirm.service'
 import { MessageService } from '../_services/message.service'
 
 @Component({
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 10
   loading = false
 
-  constructor (private messageService: MessageService) {}
+  constructor (private messageService: MessageService, private confirmService: ConfirmService) {}
 
   ngOnInit (): void {
     this.loadMessages()
@@ -37,12 +38,19 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage (id: number) {
-    this.messageService.deleteMessage(id).subscribe({
-      next: res => {
-        this.messages.splice(
-          this.messages.findIndex(m => m.id == id),
-          1
-        )
+
+    this.confirmService.confirm("Confirm Deletion", "This cannot be undone").subscribe({
+      next: ok=>{
+        if(ok){
+          this.messageService.deleteMessage(id).subscribe({
+            next: res => {
+              this.messages.splice(
+                this.messages.findIndex(m => m.id == id),
+                1
+              )
+            }
+          })
+        }
       }
     })
   }
